requirements.txt
google-api-python-client==2.126.0
google-auth==2.29.0
python-dateutil==2.9.0.post0
pytest==8.2.0
